Modificamos el metodo reaccionarObjeto de Vehiculo

+++ src/main/java/edu/fiuba/algo3/modelo/Auto.java
@@ -6,7 +6,7 @@ public class Auto extends TipoVehiculo{
 
 
     @Override
-    public float reaccionarAPozo(float cantidadDeMovimientos) {
+    public int reaccionarAPozo(int cantidadDeMovimientos) {
         return cantidadDeMovimientos + 3;
     }
     @Override
@@ -15,7 +15,7 @@ public class Auto extends TipoVehiculo{
     }
 
     @Override
-    public float reaccionarAPiquete(float cantidadDeMovimientos) {
+    public int reaccionarAPiquete(int cantidadDeMovimientos) {
         return cantidadDeMovimientos;
     }
 }
